fix: remove old debug env code and print statements

+++ mava/systems/jax/mappo/networks.py
@@ -76,7 +76,7 @@ class PPONetworks:
     ) -> Tuple[np.ndarray, Dict]:
         """TODO: Add description here."""
         actions, log_prob = self.forward_fn(self.params, observations, key)
-        actions = np.array(actions, dtype=np.int64)
+        actions = np.array(actions, dtype=np.int32)
         log_prob = np.squeeze(np.array(log_prob, dtype=np.float32))
         return actions, {"log_prob": log_prob}
